requirements.txt
pytorch_lightning==1.4.2
cmake
transformers==4.19.2
albumentations==1.1.0
bezier==2023.7.28
diffusers==0.30.3
dift==0.0.7
einops==0.4.1
face_alignment==1.4.1
ftfy==6.0.3
glfw==2.7.0
imageio==2.14.1
kornia==0.6.0
matplotlib==3.7.5
more_itertools==10.5.0
moviepy==1.0.3
natsort==8.4.0
numpy==1.24.3
omegaconf==2.1.1
opencv_python
packaging==24.1
pandas==2.0.3
Pillow==9.0.1
proglog==0.1.10
pudb==2019.2
PyOpenGL==3.1.7
PyYAML==6.0.2
regex==2024.9.11
Requests==2.32.3
scipy==1.9.1
setuptools==59.5.0
scikit-image==0.20.0
streamlit==0.73.1
tqdm==4.66.5
typing_extensions==4.12.2
torchmetrics==0.6.0
mediapipe==0.10.21
